LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-es-expected.txt

PASS navigator.language is "es-ES"
PASS successfullyParsed is true

TEST COMPLETE
